Move initial patch generation to patches::from_scratch_diff
