+++ automerge-backend/src/patches/gen_value_diff.rs
@@ -1,0 +1,19 @@
+use crate::op_handle::OpHandle;
+
+use super::PatchWorkshop;
+use automerge_protocol as amp;
+
+pub(super) fn gen_value_diff(
+    op: &OpHandle,
+    value: &amp::ScalarValue,
+    workshop: &dyn PatchWorkshop,
+) -> amp::Diff {
+    match value {
+        amp::ScalarValue::Cursor(oid) => {
+            // .expect() is okay here because we check that the cursr exists at the start of
+            // `OpSet::apply_op()`
+            amp::Diff::Cursor(workshop.find_cursor(oid).expect("missing cursor"))
+        }
+        _ => op.adjusted_value().into(),
+    }
+}
